Moved files to model and view
