+++ app/src/main/AndroidManifest.xml
@@ -11,14 +11,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.HelloWorld">
-        <activity android:name=".home.view.HomeViewActivity">
+        <activity android:name=".home.view.HomeViewActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".moredetails.fulllogic.OtherInfoWindow"
+        <activity android:name=".moredetails.home.view.OtherInfoWindow"
            android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
