+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/home/model/LastFMAPI.java
@@ -1,4 +1,4 @@
-package ayds.lisboa.songinfo.moredetails.fulllogic;
+package ayds.lisboa.songinfo.moredetails.home.model;
 
 
 import retrofit2.Call;
